+++ HaselDebug/Tabs/PacketLogs/PrepareZoningLogTab.cs
@@ -1,0 +1,67 @@
+using FFXIVClientStructs.FFXIV.Client.Game.Network;
+using FFXIVClientStructs.FFXIV.Client.Network;
+using HaselDebug.Abstracts;
+using HaselDebug.Interfaces;
+
+namespace HaselDebug.Tabs.PacketLogs;
+
+[RegisterSingleton<IPacketLogTab>(Duplicate = DuplicateStrategy.Append), AutoConstruct]
+public unsafe partial class PrepareZoningLogTab : PacketLogTab<PrepareZoningPacket>, IDisposable
+{
+    private Hook<PacketDispatcher.Delegates.HandlePrepareZoningPacket>? _hook;
+
+    public void Dispose()
+    {
+        _hook?.Dispose();
+        Clear();
+    }
+
+    private void HandlePrepareZoningPacketDetour(PrepareZoningPacket* packet, byte a2)
+    {
+        AddRecord(*packet);
+        _hook!.Original(packet, a2);
+    }
+
+    public override void Draw()
+    {
+        _hook ??= _gameInteropProvider.HookFromAddress<PacketDispatcher.Delegates.HandlePrepareZoningPacket>(PacketDispatcher.MemberFunctionPointers.HandlePrepareZoningPacket, HandlePrepareZoningPacketDetour);
+
+        var enabled = IsPacketLogEnabled;
+        if (ImGui.Checkbox("Enabled", ref enabled))
+            TogglePacketLog();
+
+        ImGui.SameLine();
+        if (ImGui.Button("Clear"))
+            Clear();
+
+        using var table = ImRaii.Table("PrepareZoningTable"u8, 2, ImGuiTableFlags.Borders | ImGuiTableFlags.ScrollY | ImGuiTableFlags.RowBg | ImGuiTableFlags.Resizable);
+        if (!table) return;
+
+        ImGui.TableSetupColumn("Time"u8, ImGuiTableColumnFlags.WidthFixed, 100);
+        ImGui.TableSetupColumn("Packet"u8, ImGuiTableColumnFlags.WidthStretch);
+        ImGui.TableSetupScrollFreeze(0, 1);
+        ImGui.TableHeadersRow();
+
+        foreach (var (i, time, packet) in Records)
+        {
+            ImGui.TableNextRow();
+            ImGui.TableNextColumn();
+            ImGui.Text(time.ToLongTimeString());
+
+            ImGui.TableNextColumn();
+            _debugRenderer.DrawPointerType(packet);
+        }
+    }
+
+    public override void EnablePacketLog()
+    {
+        _hook!.Enable();
+        IsPacketLogEnabled = _hook.IsEnabled;
+    }
+
+    public override void DisablePacketLog()
+    {
+        _hook!.Disable();
+        IsPacketLogEnabled = _hook.IsEnabled;
+    }
+}